Refactor: exports and functions heirarchy updated. refactored Color type.
