+++ src/index.ts
@@ -1,2 +1,10 @@
-export { default as colorRange } from "./color-range";
-export * from './utils'
+export { colorRange, createColor } from "./color-range";
+export {
+  isRGB,
+  isRGBString,
+  isHex,
+  stringToRGB,
+  hexToRGB,
+  arrayToRGB,
+} from "./utils";
+export * from "./types";
